// src/app/core/models/Consulta.model.ts
export class Consulta{
    motivo: string = "";
    antecedentes: string = "";
    AVLSC_OD: number = 1;
    AVLSC_OI: number = 1;
    AVLPC_OD: number = 1;
    AVLPC_OI: number = 1;
    PH_OD: number = 1;
    PH_OI: number = 1;
    LENS_OD: number = 1;
    LENS_OI: number = 1;
    BIOM_OD: number = 1;
    BIOM_OI: number = 1;
    REF_OD: number = 1;
    REF_OI: number = 1;
	QUER_OD: number = 1;
    QUER_OI: number = 1;
    PIO_OD: number = 1;
    PIO_OI: number = 1;
	TESTC_OD: number = 1;
    TESTC_OI: number = 1;
    OFTA_OD: number = 1;
    OFTA_OI: number = 1;
	KAPPA_OD: number = 1;
    KAPPA_OI: number = 1;
    fecha: string = "";
	hirschberg: string = "";
	ducciones: string = "";
    versiones: string = "";
    disposicion: string = "";
    remision: string = "";
    codigo: number = 1;
    diagnostico: string = "";
    control: string = "";
    tipoLente: string = "";
    distanciaPupilar: string = "";
    tratamiento: string = "";
    observaciones: string = "";
	idCliente: number = 1;
    idEmpleado: number = 1;
}